Controller 수정 (컴파일 에러 해결 등)

+++ src/main/java/com/bungeobbang/app/view/controller/AdminController.java
@@ -4,6 +4,7 @@ import com.bungeobbang.app.biz.order.OrderDTO;
 import com.bungeobbang.app.biz.order.OrderService;
 import com.bungeobbang.app.biz.store.StoreDTO;
 import com.bungeobbang.app.biz.store.StoreService;
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -19,6 +20,8 @@ public class AdminController {
     private OrderService orderService;
     @Autowired
     private StoreService storeService;
+    @Autowired
+    private HttpSession session;
 
     private final String FAIL_DO = "redirect:failInfo.do"; //기본 실패 처리
     private final String FAIL_URL = "failInfo2"; //실패 처리할 페이지
@@ -37,10 +40,23 @@ public class AdminController {
     private final String PAGE_REPORT_LIST = "closedStoreDeclareList"; //views 하위 신고목록 페이지
     private final String PAGE_TIPOFF_LIST = "storeDeclareList"; //views 하위 가게 제보목록 페이지
 
+    //session
+    private final String SESSION_ROLE = "userRole";
+
+    //ADMIN 구분값
+    private final String ADMIN = "ADMIN";
+
+    //가게 주문 완료처리
     @RequestMapping("/updateOrderCheck.do")
     public String updateOrderCheck(OrderDTO orderDTO, Model model) {
         log.info("log: /updateOrderCheck.do updateOrderCheck - start");
-        log.info("log: updateOrderCheck - input orderDTO: {}", orderDTO);
+        log.info("log: updateOrderCheck - param orderDTO: {}", orderDTO);
+        //만약 관리자가 아니라면
+        if(!session.getAttribute(SESSION_ROLE).equals(ADMIN)){
+            //기본 실패처리
+            log.error("log: updateOrderCheck - not admin");
+            return FAIL_DO;
+        }
         if(!orderService.update(orderDTO)){
             log.error("log: updateOrderCheck - update failed");
             model.addAttribute("msg", ORDER_UPDATE_FAIL);
@@ -51,12 +67,14 @@ public class AdminController {
         return "redirect:/loadListOrder.do";
     }
 
+    //가게 주문 리스트
     @RequestMapping("/loadListOrder.do")
     public String loadListOrder(){
         //KS 일단 보류 장바구니쪽이 정리되면 맞춰서 데이터 전달
         return PAGE_ORDER_LIST;
     }
 
+    //가게 신고 리스트
     @RequestMapping("/loadListStoreReport.do")
     public String loadListStoreReport(Model model){
         StoreDTO storeDTO = new StoreDTO();
@@ -65,7 +83,7 @@ public class AdminController {
         return PAGE_REPORT_LIST;
     }
 
-    //KS DB확인 STORE_TIP_OFF_LIST
+    //가게 제보리스트
     @RequestMapping("/loadListStoreTipOff.do")
     public String loadListStoreTipOff(Model model){
         StoreDTO storeDTO = new StoreDTO();
@@ -74,27 +92,39 @@ public class AdminController {
         return PAGE_TIPOFF_LIST;
     }
 
+    //가게 삭제
     @RequestMapping("/deleteStore.do")
     public String deleteStore(StoreDTO storeDTO, Model model){
         log.info("log: /deleteStore.do deleteStore - start");
-        log.info("log: deleteStore - input storeDTO: {}", storeDTO);
+        log.info("log: deleteStore - param storeDTO: {}", storeDTO);
+        //만약 관리자가 아니라면
+        if(!session.getAttribute(SESSION_ROLE).equals(ADMIN)){
+            //기본 실패처리
+            log.error("log: deleteStore - not admin");
+            return FAIL_DO;
+        }
+        //관리자라면 삭제 진행
         if(!storeService.delete(storeDTO)){
             log.error("log: deleteStore - delete failed");
             model.addAttribute("msg", STORE_DELETE_FAIL);
-            model.addAttribute("path", "");
+            model.addAttribute("path", "loadListStoreReport.do");
             return FAIL_URL;
         }
         log.info("log: /deleteStore.do deleteStore - end");
         return "redirect:loadListStoreReport.do";
     }
 
-    //관리자 파트에서 사용하는 기능
-
     //가게 폐점설정
     @RequestMapping("/updateStoreClose.do")
     public String updateStoreClose(StoreDTO storeDTO, Model model){
         log.info("log: /updateStoreClose.do updateStoreClose - start");
-        log.info("log: updateStoreClose - input storeDTO num : [{}]", storeDTO.getStoreNum());
+        log.info("log: updateStoreClose - param storeDTO num : [{}]", storeDTO.getStoreNum());
+        //만약 관리자가 아니라면
+        if(!session.getAttribute(SESSION_ROLE).equals(ADMIN)){
+            //기본 실패처리
+            log.error("log: updateStoreClose - not admin");
+            return FAIL_DO;
+        }
         HashMap<String, String> filterList = new HashMap<>();
         filterList.put("UPDATE_CLOSED", this.YES);
         storeDTO.setFilterList(filterList);
@@ -115,7 +145,13 @@ public class AdminController {
     @RequestMapping("/updateStoreVisible.do")
     public String updateStoreVisible(StoreDTO storeDTO, Model model){
         log.info("log: /updateStoreVisible.do updateStoreVisible - start");
-        log.info("log: updateStoreVisible - input storeDTO num : [{}]", storeDTO.getStoreNum());
+        log.info("log: updateStoreVisible - param storeDTO num : [{}]", storeDTO.getStoreNum());
+        //만약 관리자가 아니라면
+        if(!session.getAttribute(SESSION_ROLE).equals(ADMIN)){
+            //기본 실패처리
+            log.error("log: updateStoreVisible - not admin");
+            return FAIL_DO;
+        }
         HashMap<String, String> filterList = new HashMap<>();
         filterList.put("UPDATE_SECRET", this.NO);
         storeDTO.setFilterList(filterList);
